+++ src/main/java/isep/web/sakila/dao/repositories/CategoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import isep.web.sakila.jpa.entities.Category;
 
-public interface CategoryRepository extends CrudRepository<Category, Integer>
+public interface CategoryRepository extends CrudRepository<Category, Byte>
 {
 }
