+++ src/main/java/isep/web/sakila/dao/repositories/LanguageRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import isep.web.sakila.jpa.entities.Language;
 
-public interface LanguageRepository extends CrudRepository<Language, Integer>
+public interface LanguageRepository extends CrudRepository<Language, Byte>
 {
 }
